Refactor StarvingKuu
Seperate responsibility to defined classes
