+++ lib/starving_kuu.rb
@@ -1,13 +1,3 @@
-require 'yaml'
+Dir['lib/starving_kuu/*.rb'].each { |file| require_relative(file.delete_prefix('lib/')) }
 
-class StarvingKuu
-  RESTAURANT_DATA_ABSOLUTE_PATH = File.expand_path('../_data/restaurants.yml', File.dirname(__FILE__)).freeze
-
-  def initialize
-    @restaurants = YAML.load_file(RESTAURANT_DATA_ABSOLUTE_PATH)
-  end
-
-  def sample_restaurant
-    @restaurants.sample
-  end
-end
+module StarvingKuu; end
