+++ lib/starving_kuu/restaurant_selector.rb
@@ -1,0 +1,15 @@
+require 'yaml'
+
+module StarvingKuu
+  class RestaurantSelector
+    RESTAURANT_DATA_ABSOLUTE_PATH = File.expand_path('../../_data/restaurants.yml', File.dirname(__FILE__)).freeze
+
+    def initialize
+      @restaurants = YAML.load_file(RESTAURANT_DATA_ABSOLUTE_PATH)
+    end
+
+    def sample_restaurant
+      @restaurants.sample
+    end
+  end
+end
